+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+* Quest line check for story line quests.
 * Story check for each zone.
 * World map hooks for storyline quests.
 * New slash commands: `/lm`, `/loremaster`